requirements.txt
discord.py==1.0.1
requests==2.21.0
